Update Fix Expired Cluster Certificate Manual Steps.md
It’s important to note that it is okay if FabricGateway.exe is not immediately visible after applying the changes on a single node. The process will appear once you have completed all necessary steps on all seed nodes and at least three or more non-primary nodes. This is expected behavior and does not indicate an issue.